### 2025-08-05 #8751 @editorialbot
[PRE REVIEW]: pyDARTdiags: A Python package for manipulating observation sequences and calculating observation-space diagnostics for the Data Assimilation Research Testbed (DART)
**Submitting author:** @hkershaw-brown (<a href="http://orcid.org/0009-0002-3755-9978">Helen Kershaw</a>)
**Repository:** https://github.com/NCAR/pyDARTdiags
**Branch with paper.md** (empty if default branch):
**Version:** 0.6.1
