### 2025-08-05 #8751 @crvernon
:wave: @hkershaw-brown - Thank you for your submission to JOSS.  We are managing a large queue of submissions right now.  I will get you a topic editor assigned as soon as possible.
