### 2025-10-10 fd211c1
add CITATION.cff
