+++ _sources/examples/03_diagnostics/plot_evolution.rst.txt
@@ -183,7 +183,7 @@ axes object from the figure and using the `legend` method to move the legend.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.859 seconds)
+   **Total running time of the script:** (0 minutes 0.842 seconds)
 
 
 .. _sphx_glr_download_examples_03_diagnostics_plot_evolution.py:
