+++ _sources/examples/03_diagnostics/plot_rank_histogram.rst.txt
@@ -170,49 +170,49 @@ and posterior rank histograms are plotted.
     *
 
       .. image-sg:: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_001.png
-         :alt: ACARS_TEMPERATURE at Level 17500.0, Prior Rank Histogram, Posterior Rank Histogram
+         :alt: ACARS_TEMPERATURE at Level 17500.0 hPa, Prior Rank Histogram, Posterior Rank Histogram
          :srcset: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_001.png
          :class: sphx-glr-multi-img
 
     *
 
       .. image-sg:: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_002.png
-         :alt: ACARS_TEMPERATURE at Level 22500.0, Prior Rank Histogram, Posterior Rank Histogram
+         :alt: ACARS_TEMPERATURE at Level 22500.0 hPa, Prior Rank Histogram, Posterior Rank Histogram
          :srcset: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_002.png
          :class: sphx-glr-multi-img
 
     *
 
       .. image-sg:: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_003.png
-         :alt: ACARS_TEMPERATURE at Level 27500.0, Prior Rank Histogram, Posterior Rank Histogram
+         :alt: ACARS_TEMPERATURE at Level 27500.0 hPa, Prior Rank Histogram, Posterior Rank Histogram
          :srcset: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_003.png
          :class: sphx-glr-multi-img
 
     *
 
       .. image-sg:: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_004.png
-         :alt: ACARS_TEMPERATURE at Level 35000.0, Prior Rank Histogram, Posterior Rank Histogram
+         :alt: ACARS_TEMPERATURE at Level 35000.0 hPa, Prior Rank Histogram, Posterior Rank Histogram
          :srcset: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_004.png
          :class: sphx-glr-multi-img
 
     *
 
       .. image-sg:: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_005.png
-         :alt: ACARS_TEMPERATURE at Level 45000.0, Prior Rank Histogram, Posterior Rank Histogram
+         :alt: ACARS_TEMPERATURE at Level 45000.0 hPa, Prior Rank Histogram, Posterior Rank Histogram
          :srcset: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_005.png
          :class: sphx-glr-multi-img
 
     *
 
       .. image-sg:: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_006.png
-         :alt: ACARS_TEMPERATURE at Level 60000.0, Prior Rank Histogram, Posterior Rank Histogram
+         :alt: ACARS_TEMPERATURE at Level 60000.0 hPa, Prior Rank Histogram, Posterior Rank Histogram
          :srcset: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_006.png
          :class: sphx-glr-multi-img
 
     *
 
       .. image-sg:: /examples/03_diagnostics/images/sphx_glr_plot_rank_histogram_007.png
-         :alt: ACARS_TEMPERATURE at Level 77500.0, Prior Rank Histogram, Posterior Rank Histogram
[…]
