### 2025-11-17 #9439 @timothyas
Hey folks, just a bit of transparency: I'm still dealing with a bit of backlog since the US government shutdown (US federal worker here). I'm hoping to start this review later in the week, but if not, then early next week.

Looking forward to it!
