### 2025-11-27 NCAR/pyDARTdiags#108 @peanutfun
Example scripts require repository data
The documentation provides several example scripts that can be downloaded as `.py` or `.ipynb` files. However, they all try to open data that is part of the repository. Therefore, to locally execute these scripts, users have to clone the repository and place the scripts in the correct location (`/doc/userguide`). Only then, the data can actually be loaded.

It might be helpful to make the data downloadable from an online platform, such that the scripts can be executed without cloning the repository. Another option would be to include the data in the package binary and make its location available via the module import (probably related to #61), e.g.,
```python
import pydartdiags.obs_sequence.obs_sequence as obsq
from pydartdiags.data import obs_seq_final_lorenz_96

obs_seq = obsq.ObsSequence(obs_seq_final_lorenz_96)
```
---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
