### 2025-11-27 NCAR/pyDARTdiags#109 @peanutfun
Tests fail in Python 3.14
I installed `pyDARTdiags` in a Python 3.14 environment. Three tests are failing here:
```
===================================================================== short test summary info =====================================================================
FAILED tests/test_stats.py::TestPossibleVsUsed::test_possible_vs_used_by_layer - AssertionError: Attributes of DataFrame.iloc[:, 1] (column name="midpoint") are different
FAILED tests/test_stats.py::TestLayers::test_bin_by_layer_pressure - AssertionError: Attributes of DataFrame.iloc[:, 5] (column name="vlevels") are different
FAILED tests/test_stats.py::TestLayers::test_bin_by_layer_height - AssertionError: Attributes of DataFrame.iloc[:, 5] (column name="vlevels") are different
```

It looks like some rounding errors and issues with data types, probably related to some updates in pandas. I am working with Python 3.14.0, and
```
pandas==2.3.3
numpy==2.3.5
PyYAML==6.0.3
```
---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
