### 2025-11-27 NCAR/pyDARTdiags#110 @peanutfun
`setup.py` superseded by `pyproject.toml`
`setup.py` was generally superseded by the safer `pyproject.toml`. This package provides the latter. Couldn't `setup.py` be removed?

---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
