### 2025-11-27 NCAR/pyDARTdiags#111 @peanutfun
`plots.plot_rank_histogram` expects computed ranks
The documentation for `plots.plot_rank_histogram` is missing. The code apparently expects a data frame in which the ensemble member ranks have already been computed. This is not the case for `matplots.plot_rank_histogram`, where the input is a "regular" `ObsSequence.df` object.

---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
