### 2025-11-27 NCAR/pyDARTdiags#112 @peanutfun
`matplots.plot_rank_histogram` requires `levels`
The `matplots.plot_rank_histogram` requires the input of vertical levels. This seems overly restrictive, as the function cannot be used on data without vertical levels. Could the `levels` parameter be optional? Also, the function is missing from the documentation.

---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
