### 2025-11-27 NCAR/pyDARTdiags#113 @peanutfun
Simplify access to ensemble members and statistics
The `stats` submodule contains several functions for computing specific statistics on the ensemble. However, it is still cumbersome to compute "custom" statistics. To that end, it might be useful to provide a function that takes an observation sequence as input and returns, e.g., only the ensemble of a specific phase as a data frame. Then the user can simply compute statistics "on their own" and need not filter the data frame columns themselves. Would that be a useful addition to an existing workflow?

---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
