### 2025-11-27 NCAR/pyDARTdiags#114 @peanutfun
Use `ObsSequence` as function parameter instead of data frame?
The statistics and plotting functions usually expect the `ObsSequence.df` as input, and not the `ObsSequence` object itself. Why is that? In several plotting functions, users have to supply the ensemble size as parameter, although this information could also be derived from the (columns of) the data frame or from the `ObsSequence` object.

If all other functions operate on a data frame, what is the purpose of the `ObsSequence` object?

---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
