### 2025-11-27 NCAR/pyDARTdiags#115 @peanutfun
Too many public methods in `ObsSequence`
Looking at the API of [`ObsSequence`](https://ncar.github.io/pyDARTdiags/api/obs_sequence.html#obs_sequence.ObsSequence), it's very hard to discern which methods are "interesting" for me as an end user, and which are only important internally or for developers. If I see this correctly, the class does not have a single private method. It would make interacting with the object much simpler if the code would distinguish public and private methods, and potentially exclude the latter from the API docs.

---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
