### 2025-11-27 NCAR/pyDARTdiags#116 @peanutfun
Suggestions for the JOSS paper
The JOSS paper already reads well, but I have a few suggestions for improving it:
* The specifics of DART observation files (fourth paragraph) are a little too much for the Summary and would fit better at the start of the Statement of Need.
* Please cite DART when first mentioning it.
* Please add at least one citation for data assimilation in general.
* Please also mention the pandas Python package when first referring to (pandas) data frames, and provide citations for pandas, matplotlib, plotly.
* I was able to substantiate the claim that pyDARTdiags is used in CROCODILE via a [GitHub search](https://github.com/search?q=org%3ACROCODILE-CESM%20pydartdiags&type=code). A link to, or citation of, the project would help.

---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
