### 2025-11-27 NCAR/pyDARTdiags#117 @peanutfun
Provide DART workflow example
The examples in the documentation all work based off pre-computed data that is included in this repository. There are no explanations provided how this data was created, and how the manipulated datasets affect the data assimilation outcome when they are injected into the regular DART workflow again.

It would be great to provide a workflow example based on a simple DART model, like Lorenz 63 (which seems to be part of the DART tutorial). The example should show how an observation file is created with DART (just quickly, I realize that this is explained in detail in the DART docs), how it is loaded into pyDARTdiags, what manipulations one could perform, and how these then change the data assimilation outcome. The plots this package can create should ideally complement and visualize these changes.

---
*I raise this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439*
