### 2025-11-27 #9439 @peanutfun
Hi @hkershaw-brown! I have reviewed your submission, and I am happy to report that it is a concise, well documented, and well tested package with a tidy repository. I only see minor issues to resolve before we can move forward with the publication.

First of all, I have to mention that I have never used DART before. I installed it and ran the Lorenz 63 model tutorial. I then loaded the observation files written by DART into pyDARTdiags and played around with the data. Overall, it worked well. I found several minor issues that I raised in the repo.

However, there are some I would like to discuss and resolve before I can recommend your submission for publication. Most obscure for me as a DART newbie is the integration into the DART workflow. This might seem a bit trivial for somebody more involved, but I found it difficult to understand what manipulations of the observation data would make sense and how they affect the data assimilation. All data used in the extensive examples are pre-computed. It would help a lot if you added a "workflow" example with an actual DART model, in which you operate on data that is directly created by DART. You could also update one of the existing examples accordingly.

Secondly, I recommend introducing private methods and attributes for `ObsSequence`, to make its API more concise. Thirdly, the paper is well written, but I suggest moving some content from the summary to the statement of need, and adding more references.

Finally, the package would benefit a lot from simultaneous diagnostics of the observation and the model space. While the advantage over the existing Fortran/MATLAB workflow is clear as is, the current capabilities of pyDARTdiags do not replace it entirely. This would require an additional reader and analysis tools for the DART NetCDF output files. I think this is out of scope for this publication, but it would be a nice goal for future development.

---

Summarizing, I would like to see these issues resolved before my recommendation for publication. I look forward to your responses, and please also feel free to object 😁 Thank you!

- [x] https://github.com/NCAR/pyDARTdiags/issues/117
- [x] https://github.com/NCAR/pyDARTdiags/issues/115
- [x] https://github.com/NCAR/pyDARTdiags/issues/116
