### 2025-12-05 a32b869
paper: Citations for pandas, plotly, matplotlib, crocodile
DART cited at first reference.
addresses #116
