+++ paper/paper.bib
@@ -19,3 +19,87 @@
   doi       = {10.1175/2009BAMS2618.1},
   url       = {https://journals.ametsoc.org/view/journals/bams/90/9/2009bams2618_1.xml}
 }
+
+@article {DART2025,
+  author = "Mohamad El Gharamti and Helen Kershaw and Kevin Raeder and Brett Raczka and Benjamin Johnson and Marlena Smith and Jeffrey Anderson and Daniel Amrhein and Nancy Collins and Timothy Hoar and Benjamin Gaubert and Ian Grooms and Lukas Kugler",
+  title = "The Data Assimilation Research Testbed: A Robust, Scalable Software Facility with Groundbreaking Capabilities for Model-Data Integration",
+  journal = "Bulletin of the American Meteorological Society",
+  year = "2025",
+  publisher = "American Meteorological Society",
+  address = "Boston MA, USA",
+  volume = "106",
+  number = "11",
+  doi = "10.1175/BAMS-D-24-0214.1",
+  pages=      "E2328 - E2345",
+  url = "https://journals.ametsoc.org/view/journals/bams/106/11/BAMS-D-24-0214.1.xml"
+}
+
+@software{pandas,
+  author = {{The pandas development team}},
+  year = {2025},
+  doi = {10.5281/zenodo.3509134},
+  license = {BSD-3-Clause},
+  title = {{pandas-dev/pandas: Pandas}},
+  url = {https://github.com/pandas-dev/pandas}
+}
+
+@InProceedings{ mckinney-proc-scipy-2010,
+  author    = { {W}es {M}c{K}inney },
+  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  pages     = { 56 - 61 },
+  year      = { 2010 },
+  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  doi       = { 10.25080/Majora-92bf1922-00a }
+}
+
+@Article{Hunter:2007,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
+@online{plotly, 
+  author = {{Plotly Technologies Inc}.}, 
+  title = {Collaborative data science}, 
+  publisher = {Plotly Technologies Inc.}, 
+  address = {Montreal, QC}, 
+  year = {2015}, 
+  url = {https://plot.ly} 
+}
+
+@misc{CROCODILE-CESM,
+  author       = {CROCODILE-CESM},
+  title        = {CROCODILE: Regional ocean and carbon cycle modeling and data assimilation within the Community Earth System Model},
+  howpublished = {GitHub repository},
+  year         = {2025},
+  note         = {[Software]},
+  url          = {https://github.com/CROCODILE-CESM}
+}
+
+@article{Evensen:2003,
+   author    =  "G. Evensen",
+   title     =  "The ensemble Kalman filter: Theoretical formulation and practical implementation",
+   year      =  "2003",
+   journal   =  "Ocean Dyn.",
+   volume    =  "53",
+   pages     =  "343--367"
+}
+
+@article{Anderson:2009,
+   author    =  "J. L. Anderson",
[…]
