### 2025-12-05 8e35277
paper: move specifics of DART observation files to the Statement of Need
Addresses issue #116
