+++ _sources/examples/02_visualizing/index.rst.txt
@@ -25,7 +25,7 @@ Examples of visualizing observation sequence data.
   .. image:: /examples/02_visualizing/images/thumb/sphx_glr_plot_qc_hover_thumb.png
     :alt:
 
-  :ref:`sphx_glr_examples_02_visualizing_plot_qc_hover.py`
+  :doc:`/examples/02_visualizing/plot_qc_hover`
 
 .. raw:: html
 
