### 2025-12-05 36d8301
fix: citation.cff Marlee's name
fixes #107
