+++ CITATION.cff
@@ -4,8 +4,8 @@ authors:
 - family-names: "Kershaw"
   given-names: "Helen"
   orcid: "https://orcid.org/0009-0002-3755-9978"
-- family-names: "Marlee"
-  given-names: "Smith"
+- family-names: "Smith"
+  given-names: "Marlee"
   orcid: "https://orcid.org/0009-0008-7616-9060"
 - family-names: "Isaac"
   given-names: "Arseneau"
