+++ _sources/examples/02_visualizing/plot_qc_hover.rst.txt
@@ -138,7 +138,7 @@ to the columns in the dataframe to see your options.
 
     <div class="output_subarea output_html rendered_html output_result">
     <div>            <script src="https://cdnjs.cloudflare.com/ajax/libs/mathjax/2.7.5/MathJax.js?config=TeX-AMS-MML_SVG"></script><script type="text/javascript">if (window.MathJax && window.MathJax.Hub && window.MathJax.Hub.Config) {window.MathJax.Hub.Config({SVG: {font: "STIX-Web"}});}</script>                <script type="text/javascript">window.PlotlyConfig = {MathJaxConfig: 'local'};</script>
-            <script charset="utf-8" src="https://cdn.plot.ly/plotly-3.3.0.min.js" integrity="sha256-bO3dS6yCpk9aK4gUpNELtCiDeSYvGYnK7jFI58NQnHI=" crossorigin="anonymous"></script>                <div id="d97e3df6-bc52-4b81-86e4-93b01b74e40d" class="plotly-graph-div" style="height:525px; width:950px;"></div>            <script type="text/javascript">                window.PLOTLYENV=window.PLOTLYENV || {};                                if (document.getElementById("d97e3df6-bc52-4b81-86e4-93b01b74e40d")) {                    Plotly.newPlot(                        "d97e3df6-bc52-4b81-86e4-93b01b74e40d",                        [{"customdata":[[0.0,"ACARS_TEMPERATURE",230.16,1],[0.0,"ACARS_U_WIND_COMPONENT",18.4,2],[0.0,"ACARS_V_WIND_COMPONENT",1.6,3],[0.0,"ACARS_TEMPERATURE",264.16,4],[0.0,"ACARS_U_WIND_COMPONENT",11.6,5],[0.0,"ACARS_V_WIND_COMPONENT",12.4,6],[0.0,"AIRCRAFT_TEMPERATURE",232.76,7],[0.0,"AIRCRAFT_U_WIND_COMPONENT",16.7,8],[0.0,"AIRCRAFT_V_WIND_COMPONENT",3.2,9],[6.0,"ACARS_TEMPERATURE",299.46,10],[6.0,"ACARS_U_WIND_COMPONENT",0.2,11],[6.0,"ACARS_V_WIND_COMPONENT",-1.5,12],[6.0,"ACARS_TEMPERATURE",275.96,13],[6.0,"ACARS_U_WIND_COMPONENT",12.7,14],[6.0,"ACARS_V_WIND_COMPONENT",17.5,15],[0.0,"ACARS_TEMPERATURE",239.66,16],[0.0,"ACARS_U_WIND_COMPONENT",55.1,17],[0.0,"ACARS_V_WIND_COMPONENT",14.8,18],[0.0,"ACARS_TEMPERATURE",246.46,19],[0.0,"ACARS_U_WIND_COMPONENT",28.0,20],[0.0,"ACARS_V_WIND_COMPONENT",3.9,21],[0.0,"ACARS_TEMPERATURE",267.66,22],[0.0,"ACARS_U_WIND_COMPONENT",34.9,23],[0.0,"ACARS_V_WIND_COMPONENT",8.7,24],[0.0,"ACARS_TEMPERATURE",271.96,25],[0.0,"ACARS_U_WIND_C
[…]
