+++ _sources/examples/03_diagnostics/plot_rank_histogram.rst.txt
@@ -223,7 +223,7 @@ and posterior rank histograms are plotted.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 2.547 seconds)
+   **Total running time of the script:** (0 minutes 2.581 seconds)
 
 
 .. _sphx_glr_download_examples_03_diagnostics_plot_rank_histogram.py:
