+++ _sources/examples/01_manipulating/plot_join_obs_seq.rst.txt
@@ -833,7 +833,7 @@ Note, the join method will still include the required columns for the obs_seq ob
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.152 seconds)
+   **Total running time of the script:** (0 minutes 0.150 seconds)
 
 
 .. _sphx_glr_download_examples_01_manipulating_plot_join_obs_seq.py:
