+++ _sources/examples/02_visualizing/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:00.218** total execution time for 1 file **from examples/02_visualizing**:
+**00:00.220** total execution time for 1 file **from examples/02_visualizing**:
 
 .. container::
 
@@ -33,5 +33,5 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_examples_02_visualizing_plot_qc_hover.py` (``plot_qc_hover.py``)
-     - 00:00.218
+     - 00:00.220
      - 0.0
