+++ _sources/examples/01_manipulating/plot_external_fo.rst.txt
@@ -263,7 +263,7 @@ Now the 'external_FO' column has the forward operator.
  .. code-block:: none
 
 
-    ['external_FO 80 1', 0.801626086917427, 0.7577725892990687, 0.5186645227250228, 0.7388945154448633, 0.7976265442376419, 0.6153834634455718, 0.7623366159262035, 0.8079816036496986, 0.8225291878285392, 0.8457345608006379, 0.704641857958876, 0.6837452909134485, 0.71764795010239, 0.8467074633832543, 0.7268152095505278, 0.7700524208394824, 0.6195671138390445, 0.6750246414926467, 0.670144632272863, 0.6699463994908811, 0.8228164716252999, 0.9683189033861539, 0.6832720583947453, 0.6158212612242915, 0.7513438098597534, 0.7536014815941487, 0.7871772420606988, 0.7204945634290844, 0.7240463292221835, 0.734407481136278, 0.4977371905796773, 0.8171628893251345, 0.6400880262936036, 0.6115164820251049, 0.6611729677267033, 0.6535702660117496, 0.6678989364546597, 0.7918291991147506, 0.7881038518406182, 0.6752156648195603, 0.6931796521672707, 0.636700779864296, 0.7532346037093658, 0.7925761281059605, 0.7390688077364225, 0.6794765210709102, 0.7444704611576493, 0.7728121874768434, 0.6914012395324891, 0.7185003474948483, 0.810221266815396, 0.7517895321565693, 0.6602923212509448, 0.6198143773435503, 0.6649373741617062, 0.6546672826930574, 0.9231045572856037, 0.6939568010156005, 0.618055316693745, 0.785476511617708, 0.550337498128746, 0.6125674660449963, 0.7433456525791877, 0.849816735901293, 0.6901262120289096, 0.7775881845049282, 0.6923658497306122, 0.47413022624547896, 0.78084957120816, 0.7591143668445531, 0.7585222656872309, 0.557422829136486, 0.6207523229041072, 0.8325357129775041, 0.9086445322804544, 0.8264352805238587, 0.7477477774977341, 0.8883481711980152, 0.9236515088154638, 0.6731399062313453]
[…]
