### 2025-12-08 2cc796e
private methods for obs_sequence module
Reduces the public API documention, keeping docs simpler for users

addresses #115
fixes #82
