+++ src/pydartdiags/obs_sequence/obs_sequence.py
@@ -7,7 +7,7 @@ import yaml
 import struct
 
 
-def requires_assimilation_info(func):
+def _requires_assimilation_info(func):
     def wrapper(self, *args, **kwargs):
         if self.has_assimilation_info():
             return func(self, *args, **kwargs)
@@ -144,31 +144,31 @@ class ObsSequence:
             self.all_obs = []
             return
 
-        if self.is_binary(file):
-            self.header = self.read_binary_header(file)
+        if self._is_binary(file):
+            self.header = self._read_binary_header(file)
         else:
-            self.header = self.read_header(file)
+            self.header = self._read_header(file)
 
-        self.types = self.collect_obs_types(self.header)
+        self.types = self._collect_obs_types(self.header)
         self.reverse_types = {v: k for k, v in self.types.items()}
-        self.copie_names, self.n_copies = self.collect_copie_names(self.header)
-        self.n_non_qc, self.n_qc = self.num_qc_non_qc(self.header)
+        self.copie_names, self.n_copies = self._collect_copie_names(self.header)
+        self.n_non_qc, self.n_qc = self._num_qc_non_qc(self.header)
         self.non_qc_copie_names = self.copie_names[: self.n_non_qc]
         self.qc_copie_names = self.copie_names[self.n_non_qc :]
 
-        if self.is_binary(file):
-            self.seq = self.obs_binary_reader(file, self.n_copies)
+        if self._is_binary(file):
+            self.seq = self._obs_binary_reader(file, self.n_copies)
             self.loc_mod = "loc3d"  # only loc3d supported for binary, & no way to check
         else:
-            self.seq = self.obs_reader(file, self.n_copies)
+            self.seq = self._obs_reader(file, self.n_copies)
 
-        self.all_obs = self.create_all_obs()  # uses up the generator
+        self.all_obs = self._create_all_obs()  # uses up the generator
         # at this point you know if the seq is loc3d or loc1d
         if self.loc_mod == "None":
             raise ValueError(
                 "Neither 'loc3d' nor 'loc1d' could be found in the observation sequence."
             )
-        self.columns = self.column_headers()
+        self.columns = self._column_headers()
         self.df = pd.DataFrame(self.all_obs, columns=self.columns)
         if self.loc_mod == "loc3d":
             self.df["longitude"] = np.rad2deg(self.df["longitude"])
@@ -184,25 +184,25 @@ class ObsSequence:
         }
         self.df = self.df.rename(columns=rename_dict)
 
-        if self.is_binary(file):
+        if self._is_binary(file):
             # binary files do not have "OBS      X" in, so set linked list from df.
             self.update_attributes_from_df()
 
         # Replace MISSING_R8s with NaNs in posterior stats where DART_quality_control = 2
         if self.has_posterior():
-            ObsSequence.replace_qc2_nan(self.df)
+            ObsSequence._replace_qc2_nan(self.df)
 
-    def create_all_obs(self):
+    def _create_all_obs(self):
         """steps through the generator to create a
         list of all observations in the sequence
         """
         all_obs = []
         for obs in self.seq:
-            data = self.obs_to_list(obs)
+            data = self._obs_to_list(obs)
             all_obs.append(data)
         return all_obs
 
-    def obs_to_list(self, obs):
+    def _obs_to_list(self, obs):
         """put single observation into a list"""
         data = []
         data.append(int(obs[0].split()[1]))  # obs_num
@@ -236,7 +236,7 @@ class ObsSequence:
         # any observation specific obs def info is between here and the end of the list
         # can be obs_def & external forward operator
         metadata = obs[typeI + 2 : -2]
[…]
