### 2025-12-08 5deae9b
Rearranged obs_sequence module documentation
Attributes section, methods section with the methods important to a user (public methods)
