+++ src/pydartdiags/obs_sequence/obs_sequence.py
@@ -58,43 +58,8 @@ class ObsSequence:
         .. code-block:: python
 
             obs_seq = ObsSequence(file='obs_seq.final')
+            empty_obs_seq = ObsSequence(file=None)
 
-
-    Attributes:
-        df (pandas.DataFrame): The DataFrame containing the observation sequence data.
-        header (list): The header of the observation sequence.
-        copie_names (list): The names of the copies in the observation sequence.
-            Spelled 'copie' to avoid conflict with the Python built-in 'copy'.
-            Spaces are replaced with underscores in copie_names.
-        non_qc_copie_names (list): The names of the copies not including quality control,
-            e.g. observation, mean, ensemble_members
-        qc_copie_names (list): The names of the quality control copies, e.g. DART_QC
-        n_copies(int): The total number of copies in the observation sequence.
-        n_non_qc(int): The number of copies not including quality control.
-        n_qc(int): The number of quality control copies.
-        vert (dict): A dictionary mapping DART vertical coordinate types to their
-            corresponding integer values.
-
-            - undefined: 'VERTISUNDEF'
-            - surface: 'VERTISSURFACE' (value is surface elevation in meters)
-            - model level: 'VERTISLEVEL'
-            - pressure: 'VERTISPRESSURE' (in Pascals)
-            - height: 'VERTISHEIGHT' (in meters)
-            - scale height: 'VERTISSCALEHEIGHT' (unitless)
-        loc_mod (str): The location model, either 'loc3d' or 'loc1d'.
-            For 3D sphere models: latitude and longitude are in degrees in the DataFrame.
-        types (dict): Dictionary of types of observations in the observation sequence,
-            e.g. {23: 'ACARS_TEMPERATURE'},
-        reverse_types (dict): Dictionary of types with keys and values reversed, e.g
-            {'ACARS_TEMPERATURE': 23}
-        synonyms_for_obs (list): List of synonyms for the observation column in the DataFrame.
-
-
-        seq (generator): Generator of observations from the observation sequence file.
-        all_obs (list): List of all observations, each observation is a list.
-            Valid when the ObsSequence is created from a file.
-            Set to None when the ObsSequence is created from scratch or multiple
-            ObsSequences are joined.
     """
 
     vert = {
@@ -349,7 +314,9 @@ class ObsSequence:
             - A 'linked_list' column is generated to create a linked list pattern for the observations.
 
         Example:
-            obsq.write_obs_seq('obs_seq.new')
+            .. code-block:: python
+
+                obsq.write_obs_seq('obs_seq.new')
 
         """
 
@@ -432,7 +399,9 @@ class ObsSequence:
         number of observations, and constructs the header with necessary information.
 
         Example:
-        self.create_header_from_dataframe()
+            .. code-block:: python
+        
+                self.create_header_from_dataframe()
 
         """
 
@@ -998,7 +967,7 @@ class ObsSequence:
         Returns:
             A new ObsSequence object containing the combined data.
 
-        Example:
+        Examples:
             .. code-block:: python
 
                 obs_seq1 = ObsSequence(file='obs_seq1.final')
