### 2025-12-08 f63a1e2
doc: wrap for docstrings for decorated function
