### 2026-01-02 NCAR/pyDARTdiags#109 @hkershaw-brown
With 0.6.1  the tests pass:
```
(py-env314-pip) [hkershaw:pyDARTdiags](tags/0.6.1) > pytest
================================================================================================= test session starts ==================================================================================================
platform darwin -- Python 3.14.0, pytest-9.0.2, pluggy-1.6.0
rootdir: /Users/hkershaw/DART/Projects/Diagnostics/JOSS/pyDARTdiags
configfile: pyproject.toml
collected 75 items

tests/test_obs_sequence.py ............................................................                                                                                                                          [ 80%]
tests/test_stats.py ...............                                                                                                                                                                              [100%]

================================================================================================== 75 passed in 1.25s ==================================================================================================
(py-env314-pip) [hkershaw:pyDARTdiags](tags/0.6.1) >
```

I can reproduce the 3 failing stats tests with pip install pyDARTdiags (0.6.1) but running the tests with  main checked out.

```
(py-env314-pip) [hkershaw:pyDARTdiags](main) > pytest
================================================================================================= test session starts ==================================================================================================
platform darwin -- Python 3.14.0, pytest-9.0.2, pluggy-1.6.0
rootdir: /Users/hkershaw/DART/Projects/Diagnostics/JOSS/pyDARTdiags
configfile: pyproject.toml
collected 75 items

tests/test_obs_sequence.py FFFF..............................FFFFFFFF.F......F.FFFFFFFF                                                                                                                          [ 80%]
tests/test_stats.py .........F.FF..
```

Note also the tes_obs_sequence fails are from renaming the private functions.

I will bump the version on pyPI since we have put in several bug fixes on main since 0.6.1.
