+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="pydartdiags",
-    version="0.6.1",
+    version="0.6.2",
     packages=find_packages(where="src"),
     package_dir={"": "src"},
     include_package_data=True,
