+++ _sources/examples/01_manipulating/plot_external_fo.rst.txt
@@ -263,7 +263,7 @@ Now the 'external_FO' column has the forward operator.
  .. code-block:: none
 
 
-    ['external_FO 80 1', 0.7517110054043824, 0.6498791090178782, 0.9043171822707762, 0.692778093704201, 0.7957500132922433, 0.6051722414758635, 0.8017157285425647, 0.6295613892514549, 0.7567589837617561, 0.7866673912121966, 0.5643927955473703, 0.8417526337191719, 0.622860534293414, 0.7221663221038885, 0.8213641766220361, 0.895164838974474, 0.7260463033925019, 0.8924249302584985, 0.578185400277862, 0.628272146880877, 0.7789874688659252, 0.873868822989234, 0.8636122223349687, 0.7972858198627589, 0.8001830674068982, 0.8870306571954497, 0.7482797121259182, 0.6962677211471014, 0.6918064553737608, 0.6860101435654489, 0.8193750413138985, 0.8153329124665342, 0.6401953951667656, 0.6804596187132728, 0.6832443132135736, 0.7721030711397634, 0.525728911465315, 0.7652317942108533, 0.8024798660600654, 0.8186200731031822, 0.7855921861084746, 0.6860513241663825, 0.7006685489801922, 0.7240021787682762, 0.866764351932124, 0.6584263161324285, 0.893493977459152, 0.8458243378989808, 0.6260712673968548, 0.7820624309010117, 0.8172970238312098, 0.7802379163865907, 0.7904391799860853, 0.7561112880468005, 0.8730919856915005, 0.6529068683893338, 0.876351773506601, 0.7816844102634258, 0.6293803165415116, 0.7163527031064063, 0.7158099029888271, 0.8111708490843226, 0.8192877102064561, 0.5129408022725249, 0.6506378783494579, 0.8236834437122134, 0.8108444333934046, 0.739172673469396, 0.8783630998462527, 0.9026194286307028, 0.7264615830996446, 0.711309934806651, 0.8106767723305945, 0.8227227701769079, 0.8227333616590502, 0.8357878530392434, 0.7842700553330528, 0.8511645107342243, 0.9053126486761051, 0.534978898153148]
[…]
