### 2026-01-02 NCAR/pyDARTdiags#109 @hkershaw-brown
version bumped to 0.6.2 on pypi.
