### 2026-01-02 8deadd9
remove setup.py, only use pyproject.toml
