### 2026-01-05 6f8066b
feat: download data from zenodo
data.py module for managing download of example data.
The example data is too big for pypi so is stored on Zenodo

Try data directory (local installs)
Envinment variable (can be used to point to example data)
Zenodo (downloads example data to users home directory)

fixes #108
