+++ examples/03_diagnostics/plot_evolution.py
@@ -11,16 +11,13 @@ For an explanation of the statistics calculations see the :ref:`statistics` guid
 # and the matplots module for plotting.
 import pydartdiags.obs_sequence.obs_sequence as obsq
 from pydartdiags.matplots import matplots as mp
+from pydartdiags.data import get_example_data
 
 ###########################################
 # Chose an obs_seq file to read.
 # In this example, we are using "obs_seq.final.lorenz_96" which is from
 # a Lorenz 96 model run with the DART assimilation system.
-# The obs_seq.final.lorenz_96 file comes with the pyDARTdiags package
-# in the data directory, so we ``import os`` to get the path to the file.
-import os
-data_dir = os.path.join(os.getcwd(), "../..", "data")
-data_file = os.path.join(data_dir, "obs_seq.final.lorenz_96")
+data_file = get_example_data("obs_seq.final.lorenz_96")
 
 ###########################################
 # Read the obs_seq file into an obs_seq object.
