+++ docs/api/example_data.rst
@@ -1,0 +1,9 @@
+.. example_data:
+
+=====================
+module: data
+=====================
+
+.. automodule:: data
+   :members:
+   :member-order: bysource
