### 2026-01-05 42ee1a4
Adding new documentation page to walk user through workflow of using pyDARTdiags with DART, including new example to change error variance
