### 2026-01-08 7af1c11
Finishing the DART workflow doc page - removing unecessary info and instead linking to DART docs, providing direct instructions for writing Python script to half obs error variance, etc.
