+++ docs/userguide/dartworkflow.rst
@@ -11,10 +11,11 @@ This page will give you basic instructions to create an observation sequence fil
 use pyDARTdiags to perform some basic manipulations on the observation sequence, and set up
 and run data assimilation with DART using the Lorenz 63 model, which is a simple ODE model
 with only three variables. You will then analyze the results with pyDARTdiags and observe how
-the changes made to the observation sequence effected the data assimilation outcome.
+the changes made to the observation sequence affected the data assimilation outcome.
 
 This guide will go through the following steps:
 
+* Download DART
 * Set up a DART experiment with the Lorenz 63 model
 * Create an observation sequence file to be used in the assimilation
 * Run the assimilation with the original observation sequence
@@ -22,17 +23,10 @@ This guide will go through the following steps:
 * Run the assimilation with the modified observation sequence
 * Analyze and compare the results with pyDARTdiags
 
-The instructions on this page are a condensed version from what can be found in the `DART documentation <https://dart.ucar.edu/documentation/>`__.
-Specifically, the DART documentation contains:
-
-* Instructions to download DART: `Downloading DART <https://docs.dart.ucar.edu/en/latest/guide/downloading-dart.html>`__.
-* Instructions to compile DART: `Compiling DART <https://docs.dart.ucar.edu/en/latest/guide/compiling-dart.html>`__.
-* A full guide to generating DART observation files and performing DART data assimilation with the Lorenz 63 model: `Using DART with the Lorenz 63 Model <https://dart.ucar.edu/documentation/using-dart-with-the-lorenz-63-model/>`__.
-* A detailed description of the Lorenz 63 model and its relevance to data assimilation: `Lorenz 63 Model <https://dart.ucar.edu/documentation/models/lorenz-63-model/>`__.
-
-The DART documentation also provides complete tutorials and comprehensive information on data
-assimilation and setting up and running DART experiments DART with various models and configurations,
-including complete tutorials.
+The instructions on this page are a condensed version from what can be found in the
+`DART documentation <https://docs.dart.ucar.edu/en/latest/index.html>`__. The DART documentation
+also provides complete tutorials and comprehensive information on ensemble data assimilation,
+how DART works, and how to use DART with various models and configurations.
 
 
 Downloading DART
@@ -46,98 +40,35 @@ Checkout the latest release of DART:
 
 If you have forked the DART repository, replace ``NCAR`` with your Github username.
 
+See the `Downloading DART <https://docs.dart.ucar.edu/en/latest/guide/downloading-dart.html>`__
+documentation page for more information.
+
 
 Compiling DART
 --------------
 
-To build DART executables you will need to:
+To build the DART executables for Lorenz 63, follow the steps detailed in the
+`Compiling DART <https://docs.dart.ucar.edu/en/latest/guide/compiling-dart.html>`__
+documentation page.
 
-#. Create an mkmf.template with appropriate compiler and library flags.
-
-   mkmf.templates for different compilers/architectures can be found
-   in the ``DART/build_templates`` directory and have names with
-   extensions that identify the compiler, the architecture, or both. This
-   is how you inform the build process of the specifics of your system.
-
-   Copy the template that is most similar to your system into ``DART/build_templates/mkmf.template``.
-   
-   Then open ``DART/build_templates/mkmf.template`` and customize it as needed by
-   editing the lines of code shown in the following example. 
-   
-   This example uses the template ``DART/build_templates/mkmf.template.intel.linux``.
-   Note that only the relevant lines of the file are shown here. The
-   first portion of the file is a large comment block that provides
-   further advice on how to customize the *mkmf* template file if needed.
-
-   .. code-block:: text
-   
-        MPIFC = mpif90
[…]
