### 2026-01-08 a502046
Removing example for changing error variance as code was moved to doc page
