### 2026-01-09 3d8a19e
Updating the documentation page by including information on identity obs that is missing from the DART docs, since this experiment's obs seq is all identity obs. Having the user write the code to plot the rank histograms. Fixing formatting.
