+++ docs/userguide/dartworkflow.rst
@@ -28,6 +28,10 @@ The instructions on this page are a condensed version from what can be found in 
 also provides complete tutorials and comprehensive information on ensemble data assimilation,
 how DART works, and how to use DART with various models and configurations.
 
+You can learn more about the model we will be using in this example, Lorenz 63, by visiting the
+DART documentation page
+`The Lorenz 63 model and its relevance to data assimilation <https://docs.dart.ucar.edu/en/latest/guide/lorenz-63-model.html>`__.
+
 
 Downloading DART
 ----------------
@@ -87,21 +91,21 @@ namelist file by changing the ``obs_sequence_in_name`` parameter in the filter n
 Open the ``DART/models/lorenz_63/work/input.nml`` file in a text editor. Locate the filter
 section of the namelist and edit the ``obs_sequence_in_name`` parameter. Set it to the path
 of the new observation sequence file you created. See the example below.
+
 .. code-block:: fortran
 
     &filter_nml
-       ...
-       obs_sequence_in_name = 'obs_seq_with_non_default_name.out',
-       ...
+        ..
+        obs_sequence_in_name = 'obs_seq_with_non_default_name.out',
+        ..
     /
 
 Save the changes to the ``input.nml`` file.
 
 From the same ``DART/models/lorenz_63/work`` directory, run one of the following commands:
 
-``./filter``  (for non-MPI version)
+``./filter``
 
-``mpirun -n 4 ./filter``  (for MPI version with 4 processes)
 
 For more information, you can refer to the
 `Running the filter <https://docs.dart.ucar.edu/en/latest/guide/da-in-dart-with-lorenz-63.html#running-the-filter>`__
@@ -132,44 +136,50 @@ observation error variances, and write out a new observation sequence file with 
 First, ensure you have pyDARTdiags installed in your Python environment. If you haven't
 installed it yet, follow the instructions in the :ref:`installguide`.
 
-Create a new python file and open it in a text editor. Name the file
+Create a new Python file and open it in a text editor. Name the file
 ``change_error_variance.py``.
 
 Add the following blocks of code to the file. You can copy and paste the code below, making
 sure to adjust the path to your observation sequence file and file name as needed:
 
 #. Import the necessary modules.
-    .. code-block:: python
 
-        import pydartdiags.obs_sequence.obs_sequence as obsq
+   .. code-block:: python
+
+       import pydartdiags.obs_sequence.obs_sequence as obsq
 
 #. Specify the path to and name of your observation sequence file.
-    .. code-block:: python
 
-        file_name = "/path_to_your_obs_sequence_file/name_of_your_obs_sequence_file"
+   .. code-block:: python
+
+       file_name = "/path_to_your_obs_sequence_file/name_of_your_obs_sequence_file"
 
 #. Read the obs_seq file into an obs_seq object.
-    .. code-block:: python
 
-        obs_seq = obsq.ObsSequence(file_name)
+   .. code-block:: python
+
+       obs_seq = obsq.ObsSequence(file_name)
 
 #. Halve the observation error variances.
-    .. code-block:: python
 
-        obs_seq.df['error_variance'] = obs_seq.df['error_variance'] / 2.0
+   .. code-block:: python
+
+       obs_seq.df['error_variance'] = obs_seq.df['error_variance'] / 2.0
 
 #. Write out the modified observation sequence to a new file.
-    .. code-block:: python
 
-        output_file = file_name + ".half_error_variance"
-        obs_seq.write_obs_sequence(output_file)
+   .. code-block:: python
 
-        ########################## End of script ##############################
+       output_file = file_name + ".half_error_variance"
+       obs_seq.write_obs_sequence(output_file)
+
+       ########################## End of script ##############################
 
[…]
