### 2026-01-09 cc1d724
Making the levels argument optional in the plot_rank_histogram_function; adding logic to plot a skip level binning and plot a single RH if nothing for levels is passed in
