### 2026-01-12 770eb6f
Adding images of the rank histograms
