+++ docs/userguide/dartworkflow.rst
@@ -229,7 +229,7 @@ observations as assimilated as well as the ensemble forward-operator expected va
 and any quality-control values.
 
 You can now use the pyDARTdiags library to read in and analyze the observation
-space results of your DART assimilation experiment. This will be done plotting the rank
+space results of your DART assimilation experiment. In this example, we'll plot the rank
 histograms of the assimilation results.
 
 This section will guide you through the process of writing a Python script that uses
