### 2026-01-12 477fb51
Removing print statement used for testing
