### 2026-01-12 fd69c21
Formatting with black
