### 2026-01-12 b5bccdf
Updating the argument order in the call ot the plot_rank_histogram function in examples/03_diagnostics/plot_rank_histogram.py
