### 2026-01-12 57bb2a4
Removing sections on creating obs sequence, downloading, compiling, and running DART
