+++ docs/userguide/dartworkflow.rst
@@ -1,133 +1,24 @@
 .. _dartworkflow:
 
-===========================
+=================================================================
 An example of using pyDARTdiags with DART and the Lorenz 63 model
-===========================
+=================================================================
 
-This page will introduce you to the Data Assimilation Research Testbed (DART) workflow
-and demonstrate how to use pyDARTdiags to enhance your data assimilation projects.
+This page will demonstrate how to use pyDARTdiags to enhance your data assimilation projects
+with DART.
 
-This page will give you basic instructions to create an observation sequence file with DART,
-use pyDARTdiags to perform some basic manipulations on the observation sequence, and set up
-and run data assimilation with DART using the Lorenz 63 model, which is a simple ODE model
-with only three variables. You will then analyze the results with pyDARTdiags and observe how
-the changes made to the observation sequence affected the data assimilation outcome.
+We assume you have downloaded and built DART and successfully run the 
+`Lorenz 63 model <https://docs.dart.ucar.edu/en/latest/guide/da-in-dart-with-lorenz-63.html>`__.
 
-This guide will go through the following steps:
+In this guide, we'll step through:
 
-* Download DART
-* Set up a DART experiment with the Lorenz 63 model
-* Create an observation sequence file to be used in the assimilation
-* Run the assimilation with the original observation sequence
-* Modify the observation sequence using pyDARTdiags by changing the observation error variances
-* Run the assimilation with the modified observation sequence
-* Analyze and compare the results with pyDARTdiags
-
-The instructions on this page are a condensed version from what can be found in the
-`DART documentation <https://docs.dart.ucar.edu/en/latest/index.html>`__. The DART documentation
-also provides complete tutorials and comprehensive information on ensemble data assimilation,
-how DART works, and how to use DART with various models and configurations.
-
-You can learn more about the model we will be using in this example, Lorenz 63, by visiting the
-DART documentation page
-`The Lorenz 63 model and its relevance to data assimilation <https://docs.dart.ucar.edu/en/latest/guide/lorenz-63-model.html>`__.
-
-
-Downloading DART
-----------------
-
-Checkout the latest release of DART:
-
-.. code:: 
-
-   git clone https://github.com/NCAR/DART.git
-
-If you have forked the DART repository, replace ``NCAR`` with your Github username.
-
-See the `Downloading DART <https://docs.dart.ucar.edu/en/latest/guide/downloading-dart.html>`__
-documentation page for more information.
-
-
-Compiling DART
---------------
-
-To build the DART executables for Lorenz 63, follow the steps detailed in the
-`Compiling DART <https://docs.dart.ucar.edu/en/latest/guide/compiling-dart.html>`__
-documentation page.
-
-When you have completed this process, the DART executables will have been built
-in the ``DART/models/lorenz_63/work`` directory.
-If the build is successful, you will see seven executable programs
-in your lorenz_63 work directory. You can find descriptions of each program
-at the bottom of the Compiling DART documentation page linked above.
-
-
-Creating an Observation Sequence File
--------------------------------------
-
-To create an observation sequence file for the Lorenz 63 model, follow the instructions on
-the `Creating an obs_seq file of synthetic observations <https://docs.dart.ucar.edu/en/latest/guide/creating-obs-seq-synthetic.html>`__
-DART documentation page. This page will guide you through the process of generating an observation
-sequence with synthetic observations for the Lorenz 63 model using DART Fortran programs.
-
-The resulting observation sequence will be found in the ``DART/models/lorenz_63/work`` directory.
-If you use the default value for the output file name (provided in the DART documentation), the
-resulting observation sequence file will be named ``obs_seq.out``.
-
-
[…]
