### 2026-01-12 5938454
Updating description of Manipulating Observation Sequences with pyDARTdiags section
