### 2026-01-12 fc90c33
Using python interactive sessions and consoloidating the plotting of the two rank histograms into one script
