+++ docs/userguide/dartworkflow.rst
@@ -5,16 +5,16 @@ An example of using pyDARTdiags with DART and the Lorenz 63 model
 =================================================================
 
 This page will demonstrate how to use pyDARTdiags to enhance your data assimilation projects
-with DART.
+with DART by showing you how to manipulate DART observation sequences and analyze DART results.
 
 We assume you have downloaded and built DART and successfully run the 
 `Lorenz 63 model <https://docs.dart.ucar.edu/en/latest/guide/da-in-dart-with-lorenz-63.html>`__.
 
 In this guide, we'll step through:
 
-* Using pyDARTdiags to plot a rank histogram of the DART Lorenz 63 obs_seq.final
-* Using pyDARTdiags to change the observations input to filter
-* Using pyDARTdiags to plot the rank histogram for the modified observations
+* Using pyDARTdiags to change the observations input to DART's filter program
+* Using pyDARTdiags to plot rank histograms of the DART Lorenz 63 obs_seq.final files, 
+  both for the original run of filter and for the run with the modified observations
 
 
 Manipulating Observation Sequences with pyDARTdiags
@@ -28,10 +28,10 @@ file with the altered data.
 First, ensure you have pyDARTdiags installed in your Python environment. If you haven't
 installed it yet, follow the instructions in the :ref:`installguide`.
 
-Create a new Python file and open it in a text editor. Name the file
-``change_error_variance.py``.
+Start an interactive Python session by entering the command ``python`` or ``python3``
+in your terminal.
 
-Add the following blocks of code to the file. You can copy and paste the code below, making
+Enter the following lines of code one by one. You can copy and paste the code lines below, making
 sure to adjust the path to your observation sequence file and file name as needed:
 
 #. Import the necessary modules.
@@ -58,22 +58,27 @@ sure to adjust the path to your observation sequence file and file name as neede
 
        obs_seq.df['error_variance'] = obs_seq.df['error_variance'] / 2.0
 
-#. Write out the modified observation sequence to a new file.
+#. Create a new file name for the modified observation sequence.
 
    .. code-block:: python
 
        output_file = file_name + ".half_error_variance"
+
+#. Write out the modified observation sequence to a new file.
+
+   .. code-block:: python
+
        obs_seq.write_obs_sequence(output_file)
 
        ########################## End of script ##############################
 
-Save your Python script and run it.
+#. Exit the interactive session.
 
-.. code-block:: bash
+   .. code-block:: bash
 
-   python3 plot_change_error_variance.py
+       exit()
 
-This will create a new observation sequence file with the modified error variances in your
+A new observation sequence file with the modified error variances should be saved in your
 current directory. The new file will have be the name of the observation sequence file you
 specified with ``.half_error_variance`` appended to the end.
 
@@ -97,7 +102,7 @@ histograms of the assimilation results.
 
 This section will guide you through the process of writing a Python script that uses
 pyDARTdiags functions to read in the final observation sequence files and plot the rank
-histograms for the identity observations.
+histograms of the DART Lorenz 63 obs_seq.final files.
 
 The observation sequence files you created for Lorenz 63 contain only identity
 observations. An identity observation means that the physical variable being observed is 
@@ -118,18 +123,22 @@ DART recognizes a negative integer as the type code for identity observations. T
 writing the program to create the rank histograms, you will need to specify the observation type
 as a negative integer, such as ``-1``.
 
-The instructions below will now guide you through creating this program.
+Start an interactive Python session by entering the command ``python`` or ``python3``
+in your terminal.
 
-Create a new python file and open it in a text editor. Name the file ``rank_histogram.py``.
-
[…]
