### 2026-01-13 22c5055
Fixing some typos in the halve error variance code
