### 2026-01-13 8c568fa
Update language on levels for plot_rank_histogram
Co-authored-by: Helen Kershaw <20047007+hkershaw-brown@users.noreply.github.com>
