### 2026-01-13 b8efd2f
Adding reference to rank hist section of stats doc
