### 2026-01-13 d786268
Formatting with black again
