### 2026-01-14 29a8879
Having users work in the /DART/models/lorenz_63/work directory and updating obs_seq file paths
