### 2026-01-14 29c6549
Adding doc string for API guide for plot_rank_histogram
