+++ docs/userguide/dartworkflow.rst
@@ -103,7 +103,7 @@ and any quality-control values.
 
 You can now use the pyDARTdiags library to read in and analyze the observation
 space results of your DART assimilation experiment. In this example, we'll plot the
- :ref:`stats-rank-hist` of the DART Lorenz 63 obs_seq.final files.
+:ref:`stats-rank-hist` of the DART Lorenz 63 obs_seq.final files.
 
 The observation sequence files you created for Lorenz 63 contain only identity
 observations. You can read about identity observations in
