### 2026-01-14 0bc6481
doc string: identity obs message
