### 2026-01-14 14b1b21
filter dataframe for identity obs
regualar obs have the type,e.g. "RAW_STATE_VARIABLE"
identity obs are a negative number
