+++ _sources/examples/01_manipulating/plot_external_fo.rst.txt
@@ -263,7 +263,7 @@ Now the 'external_FO' column has the forward operator.
  .. code-block:: none
 
 
-    ['external_FO 80 1', 0.7672211032992364, 0.8042021675630335, 0.8125437610424033, 0.7241192287256718, 0.7511016417094752, 0.8489368059800233, 0.7533285152752548, 0.7576910781224243, 0.7619783541041808, 0.554644762060394, 0.8336516924646818, 0.7678080769322813, 0.835972186522372, 0.9070410118872893, 0.7726496376626274, 0.868303099961517, 0.7843496965975179, 0.7849304465202567, 0.6585992029798117, 0.6356865809136348, 0.8499945339435924, 0.7477705780964281, 0.799945344709683, 0.7873696969465215, 0.958576562494225, 0.7866051315990411, 0.68530099439959, 0.648716844292799, 0.7426062626239333, 0.8523788606266374, 0.5757642564411685, 0.7345269375649889, 0.7905452294042946, 0.7927919839350481, 0.5643363252360574, 0.882291962192137, 0.7405992274072248, 0.7054933219467301, 0.7118562095431965, 0.7987025979702917, 0.7705561324287336, 0.7713187258946754, 0.42389466318812147, 0.9033235817531451, 0.7908766862093771, 0.7478998721537113, 0.7795365148969217, 0.6945313244586552, 0.7979163715094597, 0.7885570812804112, 0.881088798719923, 0.8191027949468545, 0.8399389217450712, 0.7992145311814177, 0.7872840704455488, 0.7842616504005977, 0.783058800672562, 0.6654550089694646, 0.8672974628428569, 0.7053504319878251, 0.6249793738177267, 0.8213010091555337, 0.5068279486439732, 0.7728805712643673, 0.8394305980402808, 0.8493950905271338, 0.7822286050351435, 0.8116973078573541, 0.5443448696804251, 0.9188417688234443, 0.857885028645435, 0.7913595359405328, 0.7026114973919232, 0.7458158588722777, 0.649991286263848, 0.6836782963098854, 0.5950326029208834, 0.6901173954321512, 0.8205314625365687, 0.7438767445013162]
[…]
