+++ _sources/examples/03_diagnostics/plot_profiles.rst.txt
@@ -161,7 +161,7 @@ To plot only the bias, set rmse=False and totalspread=False.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.627 seconds)
+   **Total running time of the script:** (0 minutes 0.616 seconds)
 
 
 .. _sphx_glr_download_examples_03_diagnostics_plot_profiles.py:
