+++ _downloads/21f8c607b07bcc3631d8187786709535/plot_qc_poss_vs_used.py
@@ -12,15 +12,12 @@ by type.
 ###########################################
 # Import the obs_sequence module
 import pydartdiags.obs_sequence.obs_sequence as obsq
+from pydartdiags.data import get_example_data
 
 ###########################################
 # Chose an obs_seq file to read.
-# This is a small obs_seq file "obs_seq.final.ascii.small"
-# that comes with the pyDARTdiags package 
-# in the data directory, so we ``import os`` to get the path to the file
-import os
-data_dir = os.path.join(os.getcwd(), "../..", "data")
-data_file = os.path.join(data_dir, "obs_seq.final.ascii.small")
+# This is a small obs_seq file "obs_seq.final.ascii.small".
+data_file = get_example_data("obs_seq.final.ascii.small")
 
 ###########################################
 # read the obs_seq file into an obs_seq object
