+++ _downloads/c400bbdf0a68590c3a099a547513b40b/plot_remove_obs.py
@@ -10,15 +10,12 @@ observation sequence and write a new observation sequence file.
 ###########################################
 # Import the obs_sequence module
 import pydartdiags.obs_sequence.obs_sequence as obsq
+from pydartdiags.data import get_example_data
 
 ###########################################
 # Chose an obs_seq file to read.
-# In this example, we are using a small obs_seq file "obs_seq.final.medium"
-# that comes with the pyDARTdiags package 
-# in the data directory, so we ``import os`` to get the path to the file.
-import os
-data_dir = os.path.join(os.getcwd(), "../..", "data")
-data_file = os.path.join(data_dir, "obs_seq.final.ascii.medium")
+# In this example, we are using a small obs_seq file "obs_seq.final.medium".
+data_file = get_example_data("obs_seq.final.ascii.medium")
 
 ###########################################
 # Read the obs_seq file into an obs_seq object.
