+++ _downloads/90e5f29d726486f3e9470447562108a9/plot_profiles.py
@@ -12,15 +12,12 @@ the :ref:`statistics` guide.
 # and the matplots module for plotting.
 import pydartdiags.obs_sequence.obs_sequence as obsq
 from pydartdiags.matplots import matplots as mp
+from pydartdiags.data import get_example_data
 
 ###########################################
 # Chose an obs_seq file to read.
-# In this example, we are using a small obs_seq file "obs_seq.final.1000"
-# that comes with the pyDARTdiags package 
-# in the data directory, so we ``import os`` to get the path to the file.
-import os
-data_dir = os.path.join(os.getcwd(), "../..", "data")
-data_file = os.path.join(data_dir, "obs_seq.final.1000")
+# In this example, we are using a small obs_seq file "obs_seq.final.1000".
+data_file = get_example_data("obs_seq.final.1000")
 
 ###########################################
 # Read the obs_seq file into an obs_seq object.
