+++ _sources/examples/01_manipulating/plot_duplicates.rst.txt
@@ -521,7 +521,7 @@ The number of obs has been reduced by the number of duplicates
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 4.009 seconds)
+   **Total running time of the script:** (0 minutes 3.965 seconds)
 
 
 .. _sphx_glr_download_examples_01_manipulating_plot_duplicates.py:
