+++ _sources/examples/01_manipulating/plot_external_fo.rst.txt
@@ -269,7 +269,7 @@ Now the 'external_FO' column has the forward operator.
  .. code-block:: none
 
 
-    ['external_FO 80 1', 0.8946045195916809, 0.7378243371850928, 0.6496292789925879, 0.7222834373230407, 0.638038584219026, 0.9305660253209492, 0.7414003983505357, 0.6652851710838356, 0.6872648259590398, 0.8560297569519997, 0.8917348288875104, 0.474958877346913, 0.7473461003657228, 0.7065946718702951, 0.8674469451253968, 0.673801829803144, 0.9099091422355444, 0.7527141497452357, 0.6159891486597142, 0.7501019523061848, 0.6611675146058011, 0.7142085304068396, 0.5855028387534706, 0.7412553410673227, 0.6772310842893203, 0.7521974131254229, 0.749867327027255, 0.6184481012249397, 0.7224140207201042, 0.7054233954915357, 0.788038561287448, 0.7492507313728407, 0.779690995296002, 0.6977015937152012, 0.9001394804065196, 0.8611527304791966, 0.7329503976372546, 0.8659746215230109, 0.5909204287706149, 0.9159341021267764, 0.7852856880146863, 0.8183161379933145, 0.6200829950543351, 0.6482186752970088, 0.7511548524272659, 0.647726121832088, 0.8647650024003302, 0.7702447765664792, 0.9192033772725524, 0.65420406193449, 0.7516354728383386, 0.6764571338032789, 0.7922479534571502, 0.7762974557581461, 0.6299041917674386, 0.6575791579263908, 0.8411413773334471, 0.7672309214752114, 0.7183112143785043, 0.6078247250942763, 0.7323874756273449, 0.8129086277621826, 0.7932151565089193, 0.9723702068500768, 0.9368848494212771, 0.7486456516534397, 0.6731725150997888, 0.763355840364094, 0.7113740956354414, 0.6759317111365638, 0.7447329722734488, 0.7360898066566972, 0.6468142634647007, 0.731542719156891, 0.782911529819497, 0.8569618269399414, 0.6446184066419969, 0.6523093949426498, 0.8024571754833063, 0.7553928037983596]
[…]
