### 2026-01-21 NCAR/pyDARTdiags#125 @timothyas
Unable to Run through Plot Profiles Example
While running through the [Plot Profiles](https://ncar.github.io/pyDARTdiags/examples/03_diagnostics/plot_profiles.html) example notebook, specifically this block of code:

```python
type = 'ACARS_TEMPERATURE'
hPalevels = [0.0, 100.0,  150.0, 200.0, 250.0, 300.0, 400.0, 500.0, 700, 850, 925, 1000]  # hPa
levels = [i * 100 for i in hPalevels]

fig = mp.plot_profile(obs_seq, levels, type, bias=True, rmse=True, totalspread=True)
```

I encountered this issue:

```python
Cell In[5], line 5
      2 hPalevels = [0.0, 100.0,  150.0, 200.0, 250.0, 300.0, 400.0, 500.0, 700, 850, 925, 1000]  # hPa
      3 levels = [i * 100 for i in hPalevels]
----> 5 fig = mp.plot_profile(obs_seq, levels, type, bias=True, rmse=True, totalspread=True)

File [~/miniconda3/envs/dart/lib/python3.12/site-packages/pydartdiags/matplots/matplots.py:61](http://localhost:8888/lab/workspaces/auto-E/tree/Downloads/~/miniconda3/envs/dart/lib/python3.12/site-packages/pydartdiags/matplots/matplots.py#line=60), in plot_profile(obs_seq, levels, type, bias, rmse, totalspread, depth)
     58 grand = stats.grand_statistics(qc0)
     60 # add level bins to the dataframe
---> 61 stats.bin_by_layer(all_df, levels, verticalUnit=vert_unit)
     62 stats.bin_by_layer(qc0, levels, verticalUnit=vert_unit)
     64 # aggregate by layer

File [~/miniconda3/envs/dart/lib/python3.12/site-packages/pydartdiags/stats/stats.py:235](http://localhost:8888/lab/workspaces/auto-E/tree/Downloads/~/miniconda3/envs/dart/lib/python3.12/site-packages/pydartdiags/stats/stats.py#line=234), in bin_by_layer(df, levels, verticalUnit)
    231 pd.options.mode.copy_on_write = True
    232 df.loc[df["vert_unit"] == verticalUnit, "vlevels"] = pd.cut(
    233     df.loc[df["vert_unit"] == verticalUnit, "vertical"], levels, include_lowest=True
    234 )
--> 235 df.loc[:, "midpoint"] = df["vlevels"].apply(lambda x: x.mid)

File [~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/series.py:5084](http://localhost:8888/lab/workspaces/auto-E/tree/Downloads/~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/series.py#line=5083), in Series.apply(self, func, args, by_row, **kwargs)
   4960 def apply(
   4961     self,
   4962     func: AggFuncType,
   (...)   4966     **kwargs,
   4967 ) -> DataFrame | Series:
   4968     """
   4969     Invoke function on values of Series.
   4970
   (...)   5076     dtype: float64
   5077     """
   5078     return SeriesApply(
   5079         self,
   5080         func,
   5081         by_row=by_row,
   5082         args=args,
   5083         kwargs=kwargs,
-> 5084     ).apply()

File [~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/apply.py:1520](http://localhost:8888/lab/workspaces/auto-E/tree/Downloads/~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/apply.py#line=1519), in SeriesApply.apply(self)
   1517     return self.apply_compat()
   1519 # self.func is Callable
-> 1520 return self.apply_standard()

File [~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/apply.py:1578](http://localhost:8888/lab/workspaces/auto-E/tree/Downloads/~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/apply.py#line=1577), in SeriesApply.apply_standard(self)
   1576 else:
   1577     curried = func
-> 1578 mapped = obj._map_values(mapper=curried)
   1580 if len(mapped) and isinstance(mapped[0], ABCSeries):
   1581     # GH#43986 Need to do list(mapped) in order to get treated as nested
   1582     #  See also GH#25959 regarding EA support
   1583     return obj._constructor_expanddim(list(mapped), index=obj.index)

File [~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/base.py:1020](http://localhost:8888/lab/workspaces/auto-E/tree/Downloads/~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/base.py#line=1019), in IndexOpsMixin._map_values(self, mapper, na_action)
   1017 arr = self._values
   1019 if isinstance(arr, ExtensionArray):
-> 1020     return arr.map(mapper, na_action=na_action)
   1022 return algorithms.map_array(arr, mapper, na_action=na_action)

File [~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/arrays/categorical.py:1627](http://localhost:8888/lab/workspaces/auto-E/tree/Downloads/~/miniconda3/envs/dart/lib/python3.12/site-packages/pandas/core/arrays/categorical.py#line=1626), in Categorical.map(self, mapper, na_action)
   1625 na_val = np.nan
   1626 if na_action is None and has_nans:
-> 1627     na_val = mapper(np.nan) if callable(mapper) else mapper.get(np.nan, np.nan)
   1629 if new_categories.is_unique and not new_categories.hasnans and na_val is np.nan:
   1630     new_dtype = CategoricalDtype(new_categories, ordered=self.ordered)

File [~/miniconda3/envs/dart/lib/python3.12/site-packages/pydartdiags/stats/stats.py:235](http://localhost:8888/lab/workspaces/auto-E/tree/Downloads/~/miniconda3/envs/dart/lib/python3.12/site-packages/pydartdiags/stats/stats.py#line=234), in bin_by_layer.<locals>.<lambda>(x)
    231 pd.options.mode.copy_on_write = True
    232 df.loc[df["vert_unit"] == verticalUnit, "vlevels"] = pd.cut(
    233     df.loc[df["vert_unit"] == verticalUnit, "vertical"], levels, include_lowest=True
    234 )
--> 235 df.loc[:, "midpoint"] = df["vlevels"].apply(lambda x: x.mid)

AttributeError: 'float' object has no attribute 'mid'
```

---

I'm raising this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439
