### 2026-01-21 NCAR/pyDARTdiags#126 @timothyas
JOSS Paper Minor Comments
Hi folks, great work on this software package and the JOSS paper. I think the paper is really well written, short and sweet. I just have a couple of micro-comments that I'd suggest checking out.

* Line 45: Add a space between DataFrame and the Pandas citation
* Lines 59 & 60: add spaces between Matplotlib and its citation, same for Plotly
* Line 60: "facilities" should probably be "facilitates"

---

I'm raising this issue as part of my review in https://github.com/openjournals/joss-reviews/issues/9439
