### 2026-01-21 #9439 @timothyas
Hi folks! First I want to say that I am so sorry for the delay in this review! I will do my best to be more responsive - thanks for your patience.

Now to the review: great work team! I think that this is a very useful package, it's always great to see legacy FORTRAN+MATLAB pipelines get replaced with more modern solutions. To me the purpose and motivation for the package is very clear. Streamlining complex model+DA+evaluation pipelines is not straightforward, and this package does a good job of doing so. The paper is well written, and the documentation is thorough (e.g., nice work documenting all of the statistical quantities).

I do have a few comments, just a few tiny ones for the paper, and a couple of hiccups while running through the documentation (see the issues raised). But otherwise, I think this submission is looking good! Nice work!
