### 2026-01-21 99b5076
fix paper citation spacing, and misspelling of facilitates
fixes #126
