### 2026-01-21 e07cbdf
pandas dependency < 3.0.0
fixes #125
