### 2026-01-21 2ecb485
pandas < 3.0.0 for examples doc build
