+++ _sources/examples/01_manipulating/plot_external_fo.rst.txt
@@ -269,7 +269,7 @@ Now the 'external_FO' column has the forward operator.
  .. code-block:: none
 
 
-    ['external_FO 80 1', 0.7625797467948502, 0.7766293774888604, 0.7991302572744781, 0.8688107923309335, 0.5577743977077779, 0.7761277479482841, 0.7022982286555298, 0.7697072350467206, 0.615897340878128, 0.9221820757665192, 0.4915061795985205, 0.6136949604598376, 0.8564460407805191, 0.8196002806205505, 0.6263303794641316, 0.7317421747146505, 0.9625846186721556, 0.8443416088640021, 0.686982249618858, 0.8814588228176041, 0.7921029936732175, 0.8499777237378048, 0.8618484956948119, 0.7557411702017708, 0.6860375158797545, 0.7428132503847895, 0.8176883057671231, 0.8404857471298207, 0.7457630923723026, 0.5790250351916001, 0.7603400276167245, 0.9305901383815764, 0.7530373468734742, 0.7916950917279199, 0.9479709870836891, 0.6805474470551007, 0.8672272223458249, 0.6415210931714717, 0.5858058051542315, 0.7604607898529484, 0.7996727340461366, 0.7767494409855, 0.5509085298260241, 0.8860802316288509, 0.8107042200361614, 0.725182839511115, 0.916049072924709, 0.6501707173181835, 0.7869396301729276, 0.8992196230563033, 0.652869815404432, 0.5019460155108181, 0.6345242764115374, 0.6287532151773312, 0.5849337580801693, 0.5886918480210295, 0.6377919986790135, 0.5769023091015164, 0.6500655243284655, 0.7446136057017778, 0.7322252946128699, 0.7033151271913486, 0.6724736220050092, 0.5866963450405897, 0.6734483356399901, 0.8156498366582267, 0.7335069931763404, 0.8334304170815692, 0.638790962974577, 0.5976981424235102, 0.8103674678900423, 0.6353778482224091, 0.6853244177692369, 0.7420491660449041, 0.9480901578783658, 0.7502039820197489, 0.9209664302311559, 0.6544707406033023, 0.7863888004511057, 0.7268614967686488]
[…]
