### 2026-01-21 7b3bb1e
quickstart example now uses data available from pydartdiags package
updated the quickstart.rst so the output matches the obs_seq.final.1000
data

parital fix for #124
