### 2026-01-22 bcf03db
Updating the information for the first time get_example_data is introduced in the quickstart guide
