+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "pydartdiags"
-version = "0.6.3"
+version = "0.6.4"
 authors = [
   { name="Helen Kershaw", email="hkershaw@ucar.edu" },
 ]
