+++ _sources/examples/03_diagnostics/plot_qc_poss_vs_used.rst.txt
@@ -721,7 +721,7 @@ Find number of used versus possible observations
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.020 seconds)
+   **Total running time of the script:** (0 minutes 0.024 seconds)
 
 
 .. _sphx_glr_download_examples_03_diagnostics_plot_qc_poss_vs_used.py:

[…]
