### 2026-01-28 #9439 @peanutfun
@hkershaw-brown Sorry for the delay! Thanks for resolving these issues. I think the documentation and the paper improved considerably. Well done!

@mstimberg This concludes my review, and I recommend the submission for publication.
