### 2026-01-28 #9439 @mstimberg
@timothyas and @peanutfun many thanks for your time and your constructive reviews!

@hkershaw-brown with both reviewers recommending acceptance, I will now proceed with some final checks before handing over the paper to the topic editor for verification. Could you please take care of the "Author tasks" in the checklist below 👇 ? Thanks 🙏
