### 2026-01-28 74a9362
Normalize affiliation format
