### 2026-01-28 3ef4faa
Minor fix to citations, avoiding clashing parentheses
