### 2026-01-28 2f2d16f
Minor fixes for publication list (capitalization)
