+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @software{DARTcode,
   author    = {{UCAR/NSF NCAR/CISL/DAReS}},
-  title     = {The Data Assimilation Research Testbed},
+  title     = {The {Data} {Assimilation} {Research} {Testbed}},
   version   = {v11.11.2},
   date      = {2025-07},
   doi       = {10.5065/D6WQ0202},
@@ -8,7 +8,7 @@
 
 @article{DART2009,
   author    = {Jeffrey Anderson and Tim Hoar and Kevin Raeder and Hui Liu and Nancy Collins and Ryan Torn and Avelino Avellano},
-  title     = {The Data Assimilation Research Testbed: A Community Facility},
+  title     = {The {Data} {Assimilation} {Research} {Testbed}: A Community Facility},
   journal   = {Bulletin of the American Meteorological Society},
   year      = {2009},
   publisher = {American Meteorological Society},
@@ -22,7 +22,7 @@
 
 @article {DART2025,
   author = "Mohamad El Gharamti and Helen Kershaw and Kevin Raeder and Brett Raczka and Benjamin Johnson and Marlena Smith and Jeffrey Anderson and Daniel Amrhein and Nancy Collins and Timothy Hoar and Benjamin Gaubert and Ian Grooms and Lukas Kugler",
-  title = "The Data Assimilation Research Testbed: A Robust, Scalable Software Facility with Groundbreaking Capabilities for Model-Data Integration",
+  title = "The {Data} {Assimilation} {Research} {Testbed}: A Robust, Scalable Software Facility with Groundbreaking Capabilities for Model-Data Integration",
   journal = "Bulletin of the American Meteorological Society",
   year = "2025",
   publisher = "American Meteorological Society",
@@ -35,7 +35,7 @@
 }
 
 @software{pandas,
-  author = {{The pandas development team}},
+  author = {The pandas development team},
   year = {2025},
   doi = {10.5281/zenodo.3509134},
   license = {BSD-3-Clause},
@@ -44,12 +44,12 @@
 }
 
 @InProceedings{ mckinney-proc-scipy-2010,
-  author    = { {W}es {M}c{K}inney },
-  title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
-  booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+  author    = { Wes McKinney },
+  title     = { Data Structures for Statistical Computing in {P}ython },
+  booktitle = { Proceedings of the 9th {P}ython in {S}cience {C}onference (SciPy 2010)},
   pages     = { 56 - 61 },
   year      = { 2010 },
-  editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+  editor    = { St\'efan van der Walt and Jarrod Millman },
   doi       = { 10.25080/Majora-92bf1922-00a }
 }
 
@@ -88,7 +88,7 @@
 
 @article{Evensen:2003,
    author    =  "G. Evensen",
-   title     =  "The ensemble Kalman filter: Theoretical formulation and practical implementation",
+   title     =  "The ensemble {Kalman} filter: Theoretical formulation and practical implementation",
    year      =  "2003",
    journal   =  "Ocean Dyn.",
    volume    =  "53",
@@ -97,7 +97,7 @@
 
 @article{Anderson:2009,
    author    =  "J. L. Anderson",
-   title     =  "Ensemble Kalman filters for large geophysical applications",
+   title     =  "Ensemble {Kalman} filters for large geophysical applications",
    year      =  "2009",
    journal   =  "IEEE Control Syst. Mag.",
    volume    =  "29",
