### 2026-01-28 #9439 @mstimberg
@hkershaw-brown I made a PR with a few minor changes to the paper and the references (NCAR/pyDARTdiags#130), please merge if you are happy with them.
