### 2026-01-28 #9439 @mstimberg
@hkershaw-brown Apologies for all the messages – I should have checked this earlier – the bot suggests two DOIs in the comment above which look good to me, please check and add them to your BibTeX file if they match.
