+++ paper/paper.bib
@@ -35,7 +35,7 @@
 }
 
 @software{pandas,
-  author = {The pandas development team},
+  author = {{The pandas development team}},
   year = {2025},
   doi = {10.5281/zenodo.3509134},
   license = {BSD-3-Clause},
