### 2026-01-28 b1e55b4
add dois missing for 2 references
https://github.com/openjournals/joss-reviews/issues/9439#issuecomment-3812207541
