+++ paper/paper.bib
@@ -92,7 +92,8 @@
    year      =  "2003",
    journal   =  "Ocean Dyn.",
    volume    =  "53",
-   pages     =  "343--367"
+   pages     =  "343--367",
+   doi       =  "10.1007/s10236-003-0036-9"
 }
 
 @article{Anderson:2009,
@@ -101,5 +102,6 @@
    year      =  "2009",
    journal   =  "IEEE Control Syst. Mag.",
    volume    =  "29",
-   pages     =  "66--82"
+   pages     =  "66--82",
+   doi       =  "10.1109/MCS.2009.932222"
 }
