### 2026-01-28 #9439 @hkershaw-brown
> [@hkershaw-brown](https://github.com/hkershaw-brown) Apologies for all the messages – I should have checked this earlier – the bot suggests two DOIs in the comment above which look good to me, please check and add them to your BibTeX file if they match.

Thanks, dois added
