+++ _sources/examples/01_manipulating/plot_external_fo.rst.txt
@@ -269,7 +269,7 @@ Now the 'external_FO' column has the forward operator.
  .. code-block:: none
 
 
-    ['external_FO 80 1', 0.7868855836476653, 0.938230974249717, 0.7911226355664974, 0.8941686990665486, 0.6601033561107449, 0.6596735876671258, 0.6642555043376507, 0.8950578782132369, 0.6261075705528685, 0.7431859799807788, 0.7094249549877065, 1.0488026079621513, 0.6992945781353939, 0.7273262744697349, 0.5596195595572437, 0.7465108448067935, 0.9537044245067486, 0.5964317098802632, 0.7821227377130299, 0.8346105129182839, 0.6814777920648408, 0.6456506360088788, 0.73267104516247, 0.7498262059704002, 0.7725799841970677, 0.7137446615093704, 0.9610431057808317, 0.7131713202157575, 0.7935483096796119, 0.7333751069462779, 0.7674868964700426, 0.7634926138983864, 0.6559928382424635, 0.7823303974847858, 0.719888648448885, 0.8333198255132904, 0.6390258874441704, 0.7112832135494069, 0.8354237344693938, 0.7870353418831401, 0.7606665202780747, 0.7822847313609972, 0.7620167101595982, 0.5762249861316242, 0.7711672293064133, 0.5943577152799651, 0.556678643167043, 0.5910094272629411, 0.7229769297235739, 0.7387832785619706, 0.6713185272706585, 0.8112992390519117, 0.7512001413984385, 0.7459219718357352, 0.6759177662886814, 0.5490403262977086, 0.7162718959443375, 0.7457657728401551, 0.9576242034775665, 0.682888463269047, 0.8265486649718463, 0.748942254513, 0.46649544737825216, 0.6240997250920021, 0.7346351574307747, 0.7634110527768997, 0.8697996409073387, 0.7245962692942278, 0.8641364467093344, 0.7377244412520468, 0.6549169025414896, 0.8544393667989685, 0.7143659563178922, 0.7583390758950265, 0.5979667862403327, 0.6280952435936071, 0.6647075497912223, 0.614247909554217, 0.7029015634696931, 0.7285432242129283]
[…]
