### 2026-01-28 #9439 @hkershaw-brown
@mstimberg I've completed the "Additional Author Tasks After Review is Complete"

- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. **version 0.6.5**
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here. **version 0.6.5 doi: https://doi.org/10.5281/zenodo.18405764**
- [x] Make sure that the license listed for the archive is the same as the software license. ** both Apache-2.0 license**

Thank you!
