### 2026-01-29 #9439 @hkershaw-brown
> Thanks for the archive and the other information [@hkershaw-brown](https://github.com/hkershaw-brown). Could you please update the title of the archive to be the same title as the title of the paper? Also, please add the affiliations. Please note that you can simply update the metadata on Zenodo directly, there is no need for a new version.

Done! archive now has the same title as the paper.
